+++ php/widgets/term-feed/term-feed-admin.php
@@ -8,6 +8,7 @@
 ////////////////// FEED WIDGET - ADMIN //////////////////*/
 ?>
 <?php
+global $pwSiteGlobals;
 
 if ( isset( $OPTIONS[ 'title' ] ) ) {
 	  $title = $OPTIONS[ 'title' ];
@@ -16,13 +17,36 @@ if ( isset( $OPTIONS[ 'title' ] ) ) {
 	  $title = __( 'Widget', 'text_domain' );
 	}
 extract($OPTIONS);
+
+
+// Term Feed Templates
+$templates = pw_get_templates(
+	array(
+		'subdirs' => array('term-feeds'),
+		'path_type' => 'url',
+		'ext'=>'php',
+		)
+	);
+$templates = $templates['term-feeds'];
+
+// Filter which templates are available for widgets
+$show_templates = _get( $pwSiteGlobals, 'widgets.settings.term_feed.views' );
+
+// If none specified, show all
+if( empty( $show_templates ) ){
+	foreach( $templates as $key => $val ){
+		$show_templates[] = $key;
+	}
+}
+
+
 ?>
 <div class="postworld">
 	<div class="postworld-widget postworld-widget-term-feed">
 
 
 		<!-- TITLE -->
-		<label for="<?php echo $this->get_field_id( 'title' ); ?>">
+		<label class="inner" for="<?php echo $this->get_field_id( 'show_title' ); ?>">
 			<input
 				type="checkbox"
 				value="1"
@@ -30,144 +54,126 @@ extract($OPTIONS);
 				name="<?php echo $this->get_field_name('show_title'); ?>"
 				id="<?php echo $this->get_field_id('show_title'); ?>"
 				<?php if( !empty($show_title) && $show_title == '1' ){ echo 'checked="checked"'; } ?> >
-			<?php _e( 'Title:' ); ?>
+			<?php _e( 'Title' ); ?>
 		</label>
 		<input
-			class="widefat"
+			class="labeled"
 			id="<?php echo $this->get_field_id( 'title' ); ?>"
 			name="<?php echo $this->get_field_name( 'title' ); ?>"
 			type="text" value="<?php echo esc_attr( $title ); ?>" />
-
-		<div class="type-wrapper">
 		
+		<hr class="thin">
 
-			<table>
-				<!-- SELECT TAXONOMY -->
-				<tr>
-					<td>
-						<b>Taxonomy</b>
-					</td>
-					<td>
-						<select
-							name="<?php echo $this->get_field_name('taxonomy'); ?>" 
-							id="<?php echo $this->get_field_id('taxonomy'); ?>">
-						<?php
-							// Term Feed Templates
-							$taxonomy_args = array(
-								'public' => true,
-								);
-							$taxonomies = get_taxonomies( $taxonomy_args, 'objects' ); 
-							if( !empty( $taxonomies ) ){
-								foreach( $taxonomies as $tax ) {
-									$selected = '';
-									$selected = ( $tax->name == $taxonomy ) ? 'selected="selected"' : '';
-									echo '<option value="'.$tax->name.'" '.$selected.' >'.$tax->labels->name.'</option>';
-								}
-							}
-						?>
-						</select>
-					</td>
-				</tr>
+		<label
+			for="<?php echo $this->get_field_id('taxonomy'); ?>"
+			class="inner">
+			Taxonomy
+		</label>
+		<select
+			name="<?php echo $this->get_field_name('taxonomy'); ?>" 
+			id="<?php echo $this->get_field_id('taxonomy'); ?>"
+			class="labeled">
+			<?php
+				$taxonomy_args = array(
+					'public' => true,
+					);
+				$taxonomies = get_taxonomies( $taxonomy_args, 'objects' ); 
+				if( !empty( $taxonomies ) ){
+					foreach( $taxonomies as $tax ) {
+						$selected = '';
+						$selected = ( $tax->name == $taxonomy ) ? 'selected="selected"' : '';
+						echo '<option value="'.$tax->name.'" '.$selected.' >'.$tax->labels->name.'</option>';
+					}
+				}
+			?>
+		</select>
 
 
-				<!-- SELECT TERMS NUMBER -->
-				<tr>
-					<td>
-						<b>Number</b>
-					</td>
-					<td>
-						<input
-						type="number"
-						name="<?php echo $this->get_field_name('terms_number'); ?>"
-						class="short"
-						id="<?php echo $this->get_field_id('terms_number'); ?>"
-						value="<?php echo $OPTIONS['terms_number']; ?>">
-						<small>: maximum terms</small>
-					</td>
-				</tr>
+		<label
+			for="<?php echo $this->get_field_id('terms_number'); ?>"
+			class="inner">
+			Maximum Terms
+		</label>
+		<input
+			type="number"
+			name="<?php echo $this->get_field_name('terms_number'); ?>"
+			class="labeled"
+			id="<?php echo $this->get_field_id('terms_number'); ?>"
+			value="<?php echo $OPTIONS['terms_number']; ?>">
 
-				<!-- SELECT TERMS ORDER -->
-				<tr>
-					<td>
-						<b>Order</b>
-					</td>
-					<td>
-						<select
-							name="<?php echo $this->get_field_name('terms_order'); ?>" 
-							id="<?php echo $this->get_field_id('terms_order'); ?>">
-						<?php
-							$terms_order_options = array(
-								'ASC'	=>	"Ascending",
-								'DESC'	=>	"Descending",
-								);
-							foreach( $terms_order_options as $order_option => $order_name ) {
-								$selected = '';
-								$selected = ( $order_option == $OPTIONS['terms_order'] ) ? 'selected="selected"' : '';
-								echo '<option value="'.$order_option.'" '.$selected.' >'.$order_name.'</option>';
-							}
-						?>
-						</select>
-					</td>
-				</tr>
 
-				<!-- SELECT TERMS ORDER BY -->
-				<tr>
-					<td>
-						<b>Order By</b>
-					</td>
-					<td>
-						<select
-							name="<?php echo $this->get_field_name('terms_orderby'); ?>" 
-							id="<?php echo $this->get_field_id('terms_orderby'); ?>">
-						<?php
-							$terms_orderby_options = array(
-								'count'	=>	"Count (Number of posts)",
-								'name'	=>	"Name (Alphabetical)",
-								'id'	=>	"ID (Time Created)",
-								'none'	=>	"None",
-								);
-							foreach( $terms_orderby_options as $orderby_option => $orderby_name ) {
-								$selected = '';
-								$selected = ( $orderby_option == $OPTIONS['terms_orderby'] ) ? 'selected="selected"' : '';
-								echo '<option value="'.$orderby_option.'" '.$selected.' >'.$orderby_name.'</option>';
-							}
-						?>
-						</select>
-					</td>
-				</tr>
+		<label
+			for="<?php echo $this->get_field_id('terms_order'); ?>"
+			class="inner">
+			Order
+		</label>
+		<select
+			class="labeled"
+			name="<?php echo $this->get_field_name('terms_order'); ?>" 
+			id="<?php echo $this->get_field_id('terms_order'); ?>">
+			<?php
+				$terms_order_options = array(
+					'ASC'	=>	"Ascending",
+					'DESC'	=>	"Descending",
+					);
+				foreach( $terms_order_options as $order_option => $order_name ) {
+					$selected = '';
+					$selected = ( $order_option == $OPTIONS['terms_order'] ) ? 'selected="selected"' : '';
+					echo '<option value="'.$order_option.'" '.$selected.' >'.$order_name.'</option>';
+				}
+			?>
+		</select>
 
-				<!-- SELECT TEMPLATE -->
-				<tr>
-					<td>
-						<b>Template</b>
-					</td>
-					<td>
-						<select name="<?php echo $this->get_field_name('template_id'); ?>"  id="<?php echo $this->get_field_id('template_id'); ?>">
-							<?php
-							// Term Feed Templates
-							$templates = pw_get_templates(
-								array(
-									'subdirs' => array('term-feeds'),
-									'path_type' => 'url',
-									'ext'=>'php',
-									)
-								)['term-feeds'];
 
-							if( !empty( $templates ) ){
-								foreach( $templates as $template_key => $template_value ) {
-									$selected = '';
-									$selected = ( $template_id == $template_key ) ? 'selected="selected"' : '';
-									echo '<option value="'.$template_key.'" '.$selected.' >'.$template_key.'</option>';
-								}
-							}
-						?>
-						</select>
-					</td>
-				</tr>
-				
-			</table>
+		<label
+			for="<?php echo $this->get_field_id('terms_orderby'); ?>"
+			class="inner">
+			Order By
+		</label>
+		<select
+			name="<?php echo $this->get_field_name('terms_orderby'); ?>" 
+			id="<?php echo $this->get_field_id('terms_orderby'); ?>"
+			class="labeled">
+			<?php
+				$terms_orderby_options = array(
+					'count'	=>	"Count (Number of posts)",
+					'name'	=>	"Name (Alphabetical)",
+					'id'	=>	"ID (Time Created)",
+					'none'	=>	"None",
+					);
+				foreach( $terms_orderby_options as $orderby_option => $orderby_name ) {
+					$selected = '';
+					$selected = ( $orderby_option == $OPTIONS['terms_orderby'] ) ? 'selected="selected"' : '';
+					echo '<option value="'.$orderby_option.'" '.$selected.' >'.$orderby_name.'</option>';
+				}
+			?>
+		</select>
 
-	   </div>
+
+		<label
+			for="<?php echo $this->get_field_name('template_id'); ?>"
+			class="inner">
+			Template
+		</label>
+		<select
+			class="labeled"
+			name="<?php echo $this->get_field_name('template_id'); ?>"
+			id="<?php echo $this->get_field_id('template_id'); ?>">
+			<?php
+				// Show available template options
+				if( !empty( $templates ) ){
+					foreach( $templates as $template_key => $template_value ) {
+						if( !in_array( $template_key, $show_templates ) )
+							continue;
+
+						$selected = '';
+						$selected = ( $template_id == $template_key ) ? 'selected="selected"' : '';
+						echo '<option value="'.$template_key.'" '.$selected.' >'.$template_key.'</option>';
+					}
+				}
+			?>
+		</select>
+
 
 	</div>